Renamed packages and classes, included kdocs
