+++ src/main/kotlin/com/terraformation/backend/support/atlassian/request/DeleteIssueHttpRequest.kt
@@ -1,4 +1,4 @@
-package com.terraformation.backend.support.atlassian.resource
+package com.terraformation.backend.support.atlassian.request
 
 import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.statement.HttpResponse
@@ -6,7 +6,7 @@ import io.ktor.http.HttpMethod
 import io.ktor.http.path
 
 /** DELETE request to delete an issue. */
-class DeleteIssue(issueId: String) : AtlassianResource<Unit> {
+class DeleteIssueHttpRequest(issueId: String) : AtlassianHttpRequest<Unit> {
   private val path = "/rest/api/3/issue/$issueId"
   private val httpMethod = HttpMethod.Delete
