+++ src/main/kotlin/com/terraformation/backend/support/atlassian/request/ListServiceDesksHttpRequest.kt
@@ -1,16 +1,15 @@
-package com.terraformation.backend.support.atlassian.resource
+package com.terraformation.backend.support.atlassian.request
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.terraformation.backend.support.atlassian.model.ServiceRequestTypeModel
+import com.terraformation.backend.support.atlassian.model.ServiceDeskProjectModel
 import io.ktor.client.call.body
 import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.statement.HttpResponse
 import io.ktor.http.HttpMethod
 import io.ktor.http.path
 
-class ListServiceRequestTypes(serviceDeskId: Int) :
-    AtlassianResource<ListServiceRequestTypesResponse> {
-  private val path = "/rest/servicedeskapi/servicedesk/$serviceDeskId/requesttype/"
+class ListServiceDesksHttpRequest : AtlassianHttpRequest<ListServiceDesksResponse> {
+  private val path = "/rest/servicedeskapi/servicedesk/"
   private val httpMethod = HttpMethod.Get
 
   override fun buildRequest(requestBuilder: HttpRequestBuilder) {
@@ -20,10 +19,10 @@ class ListServiceRequestTypes(serviceDeskId: Int) :
     }
   }
 
-  override suspend fun parseResponse(response: HttpResponse): ListServiceRequestTypesResponse {
+  override suspend fun parseResponse(response: HttpResponse): ListServiceDesksResponse {
     return response.body()
   }
 }
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-data class ListServiceRequestTypesResponse(val values: List<ServiceRequestTypeModel>)
+data class ListServiceDesksResponse(val values: List<ServiceDeskProjectModel>)
